+++ src/Action.php
@@ -2,45 +2,31 @@
 
 namespace Telkins\LaravelPendingAction;
 
-use InvalidArgumentException;
-use Telkins\LaravelPendingAction\Contracts\Params;
+use Telkins\LaravelPendingAction\Contracts\PendingAction;
 use Telkins\LaravelPendingAction\Contracts\Action as Contract;
 
 abstract class Action implements Contract
 {
-    protected static $paramsClass;
+    protected static $pendingActionClass;
 
-    public static function prep(): Params
+    public static function prep(): PendingAction
     {
-        $paramsClass = self::getParamsClass();
-
-        return (new $paramsClass)->actionClass(static::class);
+        return self::autoPrep();
     }
 
-    private static function getParamsClass()
+    protected static function autoPrep(): PendingAction
     {
-        if (static::$paramsClass) {
-            return static::$paramsClass;
+        $pendingActionClass = self::getPendingActionClass();
+
+        return (new $pendingActionClass)->actionClass(static::class);
+    }
+
+    private static function getPendingActionClass()
+    {
+        if (static::$pendingActionClass) {
+            return static::$pendingActionClass;
         }
 
-        return static::class . 'Params';
+        return static::class . 'PendingAction';
     }
-
-    public function execute(Params $params)
-    {
-        $this->guardAgainstBadParams($params);
-
-        return $this->executeAction($params);
-    }
-
-    private function guardAgainstBadParams($params)
-    {
-        $paramsClass = self::getParamsClass();
-
-        if (! $params instanceof $paramsClass) {
-            throw new InvalidArgumentException();
-        }
-    }
-
-    abstract protected function executeAction(Params $params);
 }