+++ src/Contracts/Action.php
@@ -4,7 +4,5 @@ namespace Telkins\LaravelPendingAction\Contracts;
 
 interface Action
 {
-    public static function prep(): Params;
-
-    public function execute(Params $params);
+    public static function prep(): PendingAction;
 }