+++ tests/TestClasses/BackupDataExplicit.php
@@ -3,14 +3,13 @@
 namespace Telkins\LaravelPendingAction\Tests\TestClasses;
 
 use Telkins\LaravelPendingAction\Action;
-use Telkins\LaravelPendingAction\Contracts\Params;
 
 class BackupDataExplicit extends Action
 {
-    protected static $paramsClass = BackupDataParams::class;
+    protected static $pendingActionClass = BackupDataPendingAction::class;
 
-    protected function executeAction(Params $params)
+    public function execute(BackupDataPendingAction $pendingAction)
     {
-        // backup data for user identified by $params->username...
+        // backup data for user identified by $pendingAction->username...
     }
 }