+++ tests/TestClasses/BackupDataIdeFriendlyPendingAction.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Telkins\LaravelPendingAction\Tests\TestClasses;
+
+class BackupDataIdeFriendlyPendingAction extends BackupDataPendingAction
+{
+}