+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `laravel-pending-action` will be documented in this file
 
+## v0.2.0 - 2020-01-02
+
+- reworked action to play more nicely with IDEs
+- changed params to pending actions
+- updated tests and readme
+
 ## v0.1.0 - 2020-02-27
 
 - initial release